feat: use inheritance to improve the code base
