+++ postgrest_py/__init__.py
@@ -8,6 +8,6 @@ from postgrest_py._sync.request_builder import SyncFilterRequestBuilder  # noqa:
 from postgrest_py._sync.request_builder import SyncQueryRequestBuilder  # noqa: F401
 from postgrest_py._sync.request_builder import SyncRequestBuilder  # noqa: F401
 from postgrest_py._sync.request_builder import SyncSelectRequestBuilder  # noqa: F401
-from postgrest_py.config import DEFAULT_POSTGREST_CLIENT_HEADERS  # noqa: F401
+from postgrest_py.base_client import DEFAULT_POSTGREST_CLIENT_HEADERS  # noqa: F401
 from postgrest_py.deprecated_client import Client, PostgrestClient  # noqa: F401
 from postgrest_py.deprecated_get_request_builder import GetRequestBuilder  # noqa: F401
