--- postgrest_py/config.py
@@ -1,6 +1,0 @@
-from typing import Dict
-
-DEFAULT_POSTGREST_CLIENT_HEADERS: Dict[str, str] = {
-    "Accept": "application/json",
-    "Content-Type": "application/json",
-}
